lg_bkend/requirements.txt
fastapi
uvicorn
langchain
openai
tiktoken
pydantic
pinecone-client
pdfplumber
python-dotenv

